Fix handling of non-integer-seconds watch timeouts.
When a watch timeout is specified as something other than a multiple
of 1000ms, the fractional seconds were added to the tv_usec value of
the current time without checking if the result is greater than a
second's worth of microseconds.

This caused the timespec value used for the call to pthread_cond_timedwait
to have an invalid number of nanoseconds, which meant it would return
immediately - putting the waiting thread into an tight loop, and making
it almost impossible for the notification to be processed.